[TAN-2343] Use new GeojsonExport namespace
