+++ back/engines/commercial/custom_maps/app/controllers/custom_maps/patches/idea_custom_fields/web_api/v1/admin/idea_custom_fields_controller.rb
@@ -13,7 +13,7 @@ module CustomMaps
 
                 phase = Phase.find(params[:phase_id])
                 geojson = I18n.with_locale(current_user.locale) do
-                  CustomMaps::GeojsonGenerator.new(phase, @custom_field).generate_geojson
+                  CustomMaps::GeojsonExport::GeojsonGenerator.new(phase, @custom_field).generate_geojson
                 end
 
                 send_data geojson, type: 'application/json', filename: 'my_survey.geojson'
