--- back/engines/commercial/custom_maps/app/services/custom_maps/domicile_field_for_geojson.rb
@@ -1,9 +1,0 @@
-module CustomMaps
-  class DomicileFieldForGeojson < CustomFieldForGeojson
-    private
-
-    def option_index
-      @option_index ||= Area.all.index_by(&:id)
-    end
-  end
-end
