+++ back/engines/commercial/custom_maps/app/services/custom_maps/geojson_export/domicile_field_for_geojson.rb
@@ -1,0 +1,11 @@
+module CustomMaps
+  module GeojsonExport
+    class DomicileFieldForGeojson < CustomFieldForGeojson
+      private
+
+      def option_index
+        @option_index ||= Area.all.index_by(&:id)
+      end
+    end
+  end
+end
